# swagger.yaml
openapi: 3.0.0
info:
  title: TasteHolic API
  description: TasteHolic API 서버 문서
  version: 1.0.0
servers:
  - url: http://54.180.45.230:3000/
    description: TasteHolic Server URL
# 예시
paths:
...
  /api/v1/recipes:
    post:
      summary: 유저 레시피 등록
      description: 새로운 유저 레시피를 등록합니다.
      operationId: createUserRecipe
      tags:
        - Recipes
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - ingredients
                - recipe
                - glassType
                - status
                - abv
              properties:
                name:
                  type: string
                  example: "SampleDrink"
                ingredients:
                  type: object
                  additionalProperties:
                    type: string
                  example:
                    "라임 웨지": "3개"
                    "민트 잎": "7개"
                    "심플 시럽": "3/4 온스"
                    "라이트 럼": "2 온스"
                    "소다수": "적당량"
                recipe:
                  type: array
                  items:
                    type: string
                  example:
                    - "재료들을 함께 넣고 섞는다"
                    - "차갑게 마신다"
                glassType:
                  type: string
                  example: "Martini"
                status:
                  type: string
                  enum: [published, draft]
                  example: "published"
                tastes:
                  type: array
                  items:
                    type: string
                  example: ["Sweet", "Sour"]
                aromas:
                  type: array
                  items:
                    type: string
                  example: ["Citrus"]
                colors:
                  type: array
                  items:
                    type: string
                  example: ["#FFFFFF"]
                abv:
                  type: number
                  example: 12
      responses:
        "200":
          description: 레시피 등록 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  resultType:
                    type: string
                    example: "SUCCESS"
                  error:
                    nullable: true
                    example: null
                  success:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 11
                      userId:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: "SampleDrink"
                      imageUrl:
                        type: string
                        nullable: true
                        example: "example/test.jpg"
                      ingredients:
                        type: object
                        additionalProperties:
                          type: string
                        example:
                          "소다수": "적당량"
                          "민트 잎": "5개"
                          "라이트 럼": "3 온스"
                          "라임 웨지": "2개"
                          "심플 시럽": "1 온스"
                      recipe:
                        type: array
                        items:
                          type: string
                        example:
                          - "재료들을 함께 넣고 흔든다"
                          - "얼음과 함께 차갑게 마신다"
                      glassType:
                        type: string
                        example: "Old Fashioned"
                      createdAt:
                        type: string
                        format: date-time
                        example: "2025-02-02T14:00:50.000Z"
                      updatedAt:
                        type: string
                        format: date-time
                        example: "2025-02-02T15:30:20.000Z"
                      status:
                        type: string
                        enum: [published, draft]
                        example: "published"
                      tastes:
                        type: array
                        items:
                          type: string
                        example: ["Sweet", "Bitter"]
                      aromas:
                        type: array
                        items:
                          type: string
                        example: ["Woody"]
                      colors:
                        type: array
                        items:
                          type: string
                        nullable: true
                        example: ["#FFFFFF"]
                      likes:
                        type: integer
                        example: 0
                      abv:
                        type: number
                        example: 15
                      views:
                        type: integer
                        example: 10
    get:
      summary: 레시피 리스트 조회
      description: 레시피 목록을 조회하는 API입니다.
      operationId: getRecipeList
      tags:
        - Recipes
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [user, zero, high, fruity, under2, fav]
          description: "레시피 필터링 타입 (user: 유저 레시피, zero: 무알콜, high: 도수 높은 칵테일, fruity: 과일 향 레시피, under2: 재료 2개 이하, fav: 내가 좋아요한한 레시피)"
        - name: cursor
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: "페이지네이션을 위한 커서 (이전 응답의 nextCursor 값) / 내가 등록한 레시피 제외"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: "가져올 레시피 개수 (기본값: 10) / 내가 등록한 레시피 제외"
      responses:
        "200":
          description: 레시피 리스트 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  recipes:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 5
                        imageUrl:
                          type: string
                          example: "https://images.cocktailflow.com/v1/cocktail/w_300,h_540/cocktail_vodka_stinger-1.png"
                        nameKor:
                          type: string
                          example: "보드카 스팅거"
                        ingredients:
                          type: object
                          example:
                            "Vodka": "1 oz"
                            "White Crème de Menthe": "1 oz"
                        likes:
                          type: integer
                          example: 10
                        views:
                          type: integer
                          example: 100
                        type:
                          type: string
                          example: "cocktail"
                        myBar:
                          type: boolean
                          example: true
                  nextCursor:
                    type: integer
                    nullable: true
                    example: 4
        "400":
          description: 잘못된 요청 (잘못된 타입 입력)
          content:
            application/json:
              schema:
                type: object
                properties:
                  resultType:
                    type: string
                    example: "FAIL"
                  error:
                    type: object
                    properties:
                      errorCode:
                        type: string
                        example: "R103"
                      reason:
                        type: string
                        example: "타입 값이 잘못되었습니다. (user, zero, high, fruity, under2만 가능)"
                      data:
                        nullable: true
                        example: null
                  success:
                    nullable: true
                    example: null
...
  /users/recipes/fav:
    get:
      tags:
        - Recipes
      summary: "사용자가 좋아요한 레시피 목록 조회"
      description: "사용자가 좋아요한 레시피 목록을 페이징 방식으로 가져옵니다."
      operationId: getFavRecipes
      security:
        - bearerAuth: []
      parameters:
        - name: cursor
          in: query
          required: false
          description: "이전 요청에서 받은 nextCursor 값을 사용하여 페이지네이션"
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          description: "가져올 레시피 개수 (기본값: 10)"
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: "사용자가 좋아요한 레시피 목록을 반환합니다."
          content:
            application/json:
              schema:
                type: object
                properties:
                  recipes:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 20
                        name:
                          type: string
                          example: "SampleDrink2"
                        ingredients:
                          type: object
                          additionalProperties:
                            type: string
                          example:
                            "소다수": "적당량"
                            "민트 잎": "7개"
                            "라이트 럼": "2 온스"
                            "라임 웨지": "3개"
                            "심플 시럽": "3/4 온스"
                        likes:
                          type: integer
                          example: 1
                        views:
                          type: integer
                          example: 0
                        type:
                          type: string
                          enum: ["user", "cocktail"]
                          example: "user"
                  nextCursor:
                    type: integer
                    nullable: true
                    example: null
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT